Moved turtle stuff out of state.h/c into new files
