+++ src/state.h
@@ -6,6 +6,7 @@
 #include <stdint.h>
 
 #include "command.h"
+#include "turtle.h"
 
 #define STACK_SIZE 8
 #define CV_COUNT 4
@@ -21,56 +22,12 @@
 #define EXEC_DEPTH 8
 #define WHILE_DEPTH 10000
 
-typedef enum {
-    TT_SCRIPT_1=0,
-    TT_SCRIPT_2,
-    TT_SCRIPT_3,
-    TT_SCRIPT_4,
-    TT_SCRIPT_5,
-    TT_SCRIPT_6,
-    TT_SCRIPT_7,
-    TT_SCRIPT_8,
-    METRO_SCRIPT,
-    INIT_SCRIPT,
-    TEMP_SCRIPT
-} script_number_t;
-
 #define METRO_MIN_MS 25
 #define METRO_MIN_UNSUPPORTED_MS 2
 
 ////////////////////////////////////////////////////////////////////////////////
 // SCENE STATE /////////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////
-
-typedef struct {
-    int32_t x;         // higher resolution to permit fixed-point math
-    int32_t y;
-} turtle_position_t;
-
-typedef struct {
-    uint8_t x1;
-    uint8_t y1;
-    uint8_t x2;
-    uint8_t y2;
-} turtle_fence_t;
-
-typedef enum {
-    TURTLE_WRAP,
-    TURTLE_BUMP,
-    TURTLE_BOUNCE
-} turtle_mode_t;
-
-typedef struct {
-    turtle_position_t   position;
-    turtle_position_t   last;
-    turtle_fence_t      fence;
-    turtle_mode_t       mode;
-    uint16_t            heading;
-    int16_t             speed;
-    script_number_t     script_number;
-    bool                stepping;
-    bool                stepped;
-} scene_turtle_t;
 
 typedef struct {
     int16_t a;
@@ -200,49 +157,10 @@ scene_script_t *ss_scripts_ptr(scene_state_t *ss);
 size_t ss_scripts_size(void);
 int16_t ss_get_script_last(scene_state_t *ss, script_number_t idx);
 void ss_update_script_last(scene_state_t *ss, script_number_t idx);
-
-////////////////////////////////////////////////////////////////////////////////
-// TURTLE STATE :) /////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////////////////////////////////
-
-/*
-void     scene_set_turtle(scene_state_t*, scene_turtle_t*);
-scene_turtle_t*
-         scene_get_turtle(scene_state_t*);
-*/
-void     turtle_normalize_position(scene_turtle_t*, turtle_position_t*,
-                                   turtle_mode_t);
-void     turtle_resolve_position(scene_turtle_t*, turtle_position_t*,
-                                                  turtle_position_t*);
-uint8_t  turtle_get_x(scene_turtle_t*);
-void     turtle_set_x(scene_turtle_t*, int16_t);
-uint8_t  turtle_get_y(scene_turtle_t*);
-void     turtle_set_y(scene_turtle_t*, int16_t);
-void     turtle_move(scene_turtle_t*, int16_t, int16_t);
-void     turtle_step(scene_turtle_t*);
-int16_t  turtle_get(scene_state_t*, scene_turtle_t*);
-void     turtle_set(scene_state_t*, scene_turtle_t*, int16_t);
-/*
-void     turtle_set_home(scene_turtle_t*, int16_t, int16_t);
-uint8_t  turtle_get_home_x(scene_turtle_t*);
-uint8_t  turtle_get_home_y(scene_turtle_t*);
-void     turtle_set_home_x(scene_turtle_t*, int16_t);
-void     turtle_set_home_y(scene_turtle_t*, int16_t);
-*/
-turtle_fence_t *
-         turtle_get_fence(scene_turtle_t*);
-void     turtle_correct_fence(scene_turtle_t*);
-void     turtle_set_fence(scene_turtle_t*, int16_t, int16_t, int16_t, int16_t);
-turtle_mode_t
-         turtle_get_mode(scene_turtle_t*);
-void     turtle_set_mode(scene_turtle_t*, turtle_mode_t);
-uint16_t turtle_get_heading(scene_turtle_t*);
-void     turtle_set_heading(scene_turtle_t*, int16_t);
-int16_t  turtle_get_speed(scene_turtle_t*);
-void     turtle_set_speed(scene_turtle_t*, int16_t);
-script_number_t
-         turtle_get_script(scene_turtle_t*);
-void     turtle_set_script(scene_turtle_t*, script_number_t);
+scene_turtle_t *ss_turtle_get(scene_state_t*);
+void     ss_turtle_set(scene_state_t*, scene_turtle_t*);
+int16_t  ss_turtle_get_val(scene_state_t*, scene_turtle_t*);
+void     ss_turtle_set_val(scene_state_t*, scene_turtle_t*, int16_t);
 
 ////////////////////////////////////////////////////////////////////////////////
 // EXEC STATE //////////////////////////////////////////////////////////////////
